+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 - export to any kind of image format which are supported by Qt (JPEG, WEBP, etc...)
 - option to limit the range of the exported items
 - support TMI metadata
+- support upscaled tilesets
+- added usage report for frames and subtiles of tilesets
 
 ### Fixed
 - memory leaks